backend/agents/requirements.txt
# Sentinel Agent Dependencies
pymongo>=4.6.0
tensorflow
Pillow>=10.0.0
requests>=2.31.0
numpy>=1.24.0
python-dotenv>=1.0.0
ortools